Store dynamic settings for LTI

+++ app/models/microsoft_sync/user_mapping.rb
@@ -18,6 +18,8 @@
 # with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 
+require_dependency "microsoft_sync"
+
 #
 # See MicrosoftSync::Group for more info on Microsoft sync. This model is
 # essentially a cache between a Canvas user, and their Microsoft AAD object ID
@@ -87,7 +89,7 @@ class MicrosoftSync::UserMapping < ActiveRecord::Base
   end
 
   def self.user_ids_without_mappings(user_ids, root_account_id)
-    existing_mappings = where(root_account_id:, user_id: user_ids)
+    existing_mappings = where(root_account_id: root_account_id, user_id: user_ids)
     user_ids - existing_mappings.pluck(:user_id)
   end
 
@@ -112,10 +114,8 @@ class MicrosoftSync::UserMapping < ActiveRecord::Base
     records = user_id_to_aad_hash.map do |user_id, aad_id|
       {
         root_account_id: root_account.id,
-        created_at: now,
-        updated_at: now,
-        user_id:,
-        aad_id:,
+        created_at: now, updated_at: now,
+        user_id: user_id, aad_id: aad_id,
         needs_updating: false
       }
     end
@@ -180,7 +180,7 @@ class MicrosoftSync::UserMapping < ActiveRecord::Base
                       acct.settings[:microsoft_sync_login_attribute] == "email"
 
           GuardRail.activate(:primary) do
-            where(user:, root_account: acct)
+            where(user: user, root_account: acct)
               .update_all(updated_at: Time.now, needs_updating: true)
           end
         end
@@ -190,7 +190,7 @@ class MicrosoftSync::UserMapping < ActiveRecord::Base
 
   def self.delete_if_needs_updating(root_account_id, user_ids)
     user_ids.each_slice(1000) do |batch|
-      where(root_account_id:, user_id: batch, needs_updating: true).delete_all
+      where(root_account_id: root_account_id, user_id: batch, needs_updating: true).delete_all
     end
   end
 end
