+++ lib/canvas/migration/archive.rb
@@ -124,7 +124,7 @@ module Canvas::Migration
 
       Rails.logger.debug "Extracting #{path} to #{unzipped_file_path}"
 
-      warnings = CanvasUnzip.extract_archive(path, unzipped_file_path, nested_dir:)
+      warnings = CanvasUnzip.extract_archive(path, unzipped_file_path, nested_dir: nested_dir)
       @unzipped = true
       unless warnings.empty?
         diagnostic_text = ""
@@ -168,7 +168,9 @@ module Canvas::Migration
     end
 
     def delete_unzipped_file
-      FileUtils.rm_rf(unzipped_file_path)
+      if File.exist?(unzipped_file_path)
+        FileUtils.rm_rf(unzipped_file_path)
+      end
     end
 
     def add_warning(warning)
